+++ src/components/namespace/index.test.ts
@@ -14,6 +14,10 @@ test.each([
 
 const validParams: Params = {
   domain: "",
+  git: {
+    branch: "",
+    remote: "",
+  },
   namespace: { name: "sample-42-my-test" },
   subdomain: "",
 };