+++ src/utils/createIngress.ts
@@ -1,22 +1,25 @@
 import { Ingress } from "kubernetes-models/extensions/v1beta1/Ingress";
 
-interface IngressConfig {
+export interface IngressConfig {
   name: string;
   host: string;
   serviceName: string;
-  servicePort: string;
+  servicePort: number;
   secretName?: string;
 }
 
 export default (params: IngressConfig): Ingress => {
   const host = params.host;
+  const annotations: Record<string, string> = {
+    "kubernetes.io/ingress.class": "nginx",
+  };
+  if (process.env.PRODUCTION) {
+    annotations["certmanager.k8s.io/cluster-issuer"] = "letsencrypt-prod";
+    annotations["kubernetes.io/tls-acme"] = "true";
+  }
   return new Ingress({
     metadata: {
-      annotations: {
-        //      "certmanager.k8s.io/cluster-issuer": "letsencrypt-prod",
-        "kubernetes.io/ingress.class": "nginx",
-        //    "kubernetes.io/tls-acme": "true",
-      },
+      annotations,
       labels: {
         app: params.name,
       },