+++ src/utils/createNamespace.ts
@@ -1,24 +1,24 @@
 import { Namespace as K8SNamespace } from "kubernetes-models/v1/Namespace";
 
 import gitlab from "../environments/gitlab";
+import { merge } from "./merge";
 
 export const createNamespace = (
   // eslint-disable-next-line @typescript-eslint/ban-types
   config: object = {}
 ): K8SNamespace => {
   // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-  const envParams = {
-    ...gitlab(process.env),
-    ...config, // create options
-  };
+  const envParams = merge(
+    gitlab(process.env),
+    config // create options
+  );
   const namespace = new K8SNamespace({
     metadata: {
       annotations: {
         "field.cattle.io/creatorId": "gitlab",
-        "field.cattle.io/projectId": "c-f8qps:p-cxcnl",
-        "git/branch": "master",
-        "git/remote":
-          "https://gitlab-ci-token:x-yTUFRxUk1dxH6n8cvQ@gitlab.factory.social.gouv.fr/SocialGouv/cdtn-api.git",
+        "field.cattle.io/projectId": envParams.rancherId ?? "",
+        "git/branch": envParams.git.branch ?? "",
+        "git/remote": envParams.git.remote ?? "",
         ...envParams.annotations,
       },
       labels: envParams.labels,