+++ templates/simple/components/hasura.ts
@@ -1,0 +1,8 @@
+import env from "@kosko/env";
+import { create } from "@socialgouv/kosko-charts/components/hasura";
+
+const manifests = create({
+  env,
+});
+
+export default manifests;