+++ CHANGELOG.md
@@ -1,3 +1,198 @@
+# [2.0.0-alpha.31](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.30...v2.0.0-alpha.31) (2020-07-30)
+
+
+### Bug Fixes
+
+* **ingress:** add prod annotations ([a309eb3](https://github.com/SocialGouv/kosko-charts/commit/a309eb38ec7347460735e20f946f5b72a669e7ef))
+
+# [2.0.0-alpha.30](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.29...v2.0.0-alpha.30) (2020-07-30)
+
+
+### Bug Fixes
+
+* **utils:** ensure to use the azure-pg-user on tags ([638c457](https://github.com/SocialGouv/kosko-charts/commit/638c457a7816654b7ec1a8ca32ed58c806c03ce4))
+
+# [2.0.0-alpha.29](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.28...v2.0.0-alpha.29) (2020-07-30)
+
+
+### Bug Fixes
+
+* **hasura:** use azure fix pg user on tags ([fda47d0](https://github.com/SocialGouv/kosko-charts/commit/fda47d04b0abfd75bc309bc6d0412b2cc9851d24))
+
+# [2.0.0-alpha.28](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.27...v2.0.0-alpha.28) (2020-07-29)
+
+
+### Bug Fixes
+
+* tmp fix addWaitForPostgres ([d5eef8e](https://github.com/SocialGouv/kosko-charts/commit/d5eef8e0fd153225951946ce4d4977f28fc3dc9d))
+
+# [2.0.0-alpha.27](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.26...v2.0.0-alpha.27) (2020-07-29)
+
+
+### Bug Fixes
+
+* use pg.sealed-secret.yaml ([c08771e](https://github.com/SocialGouv/kosko-charts/commit/c08771e3fca1dcb068aeff8eeec7c1e3a405b840))
+
+# [2.0.0-alpha.26](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.25...v2.0.0-alpha.26) (2020-07-29)
+
+
+### Reverts
+
+* Revert "e2e(templates): update snapshots" ([8d33664](https://github.com/SocialGouv/kosko-charts/commit/8d3366441fd039c8d3ec9a75cd43b99480dcede7))
+
+# [2.0.0-alpha.25](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.24...v2.0.0-alpha.25) (2020-07-29)
+
+
+### Bug Fixes
+
+* **k8s:** preprod namespace name ([f9ea78f](https://github.com/SocialGouv/kosko-charts/commit/f9ea78f366ad7d86e0940c22478ed3868a95aba2))
+
+
+### Features
+
+* **k8s:** add subdomain prefix ([88046e2](https://github.com/SocialGouv/kosko-charts/commit/88046e2336e64a12042baa2a58f7a8ede0e464b2))
+
+# [2.0.0-alpha.24](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.23...v2.0.0-alpha.24) (2020-07-29)
+
+
+### Bug Fixes
+
+* type azure-pg ([fca477b](https://github.com/SocialGouv/kosko-charts/commit/fca477b126e7bb06011a045015e1fcc1e4108395))
+
+# [2.0.0-alpha.23](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.22...v2.0.0-alpha.23) (2020-07-29)
+
+
+### Bug Fixes
+
+* make hasura/config optional ([e4d4958](https://github.com/SocialGouv/kosko-charts/commit/e4d4958ba720b427f89403a715688dfdb43ee7dd))
+
+# [2.0.0-alpha.22](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.21...v2.0.0-alpha.22) (2020-07-29)
+
+
+### Features
+
+* **app:** allow optional config ([31d9b50](https://github.com/SocialGouv/kosko-charts/commit/31d9b501eeb2ecd491ecb415c032a4b307050511))
+
+# [2.0.0-alpha.21](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.20...v2.0.0-alpha.21) (2020-07-29)
+
+
+### Bug Fixes
+
+* **app:** forbid the use of containerPort in external deployment config ([6e4592c](https://github.com/SocialGouv/kosko-charts/commit/6e4592c3eb1007dccf096260aac6f46e557d09e2))
+
+# [2.0.0-alpha.20](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.19...v2.0.0-alpha.20) (2020-07-29)
+
+
+### Bug Fixes
+
+* **wait-for-pg:** use dummy psql ([8c4f0c7](https://github.com/SocialGouv/kosko-charts/commit/8c4f0c736e460182f8cc5d34f2925c3cbb0596bb))
+
+# [2.0.0-alpha.19](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.18...v2.0.0-alpha.19) (2020-07-28)
+
+
+### Bug Fixes
+
+* tests ([4f83185](https://github.com/SocialGouv/kosko-charts/commit/4f83185f5057a93c7a85c611f6d943f46c6c8e52))
+
+
+### Features
+
+* **app:** merge containers ([731fdd5](https://github.com/SocialGouv/kosko-charts/commit/731fdd56a0fe5f614a0915d05e76fe2d697cd45c))
+
+# [2.0.0-alpha.18](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.17...v2.0.0-alpha.18) (2020-07-28)
+
+
+### Bug Fixes
+
+* test pg script ([731630e](https://github.com/SocialGouv/kosko-charts/commit/731630ed615b51035e005734e6d7c8da3a98c380))
+* tests ([75de4c3](https://github.com/SocialGouv/kosko-charts/commit/75de4c32bc6491445557e4eb6305583554acc719))
+
+# [2.0.0-alpha.17](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.16...v2.0.0-alpha.17) (2020-07-28)
+
+
+### Bug Fixes
+
+* typo ([60ed9a1](https://github.com/SocialGouv/kosko-charts/commit/60ed9a1c42824f6dd58217a4dc5606adf3eede61))
+* typo ([875b48d](https://github.com/SocialGouv/kosko-charts/commit/875b48d3ac16dcd2699ea6666fe87b39bd411943))
+
+# [2.0.0-alpha.16](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.15...v2.0.0-alpha.16) (2020-07-28)
+
+
+### Features
+
+* add addWaitForPostgres and use in hasura component ([f8181ff](https://github.com/SocialGouv/kosko-charts/commit/f8181ffae293d06819d4d59e3af1f411c8475392))
+
+# [2.0.0-alpha.15](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.14...v2.0.0-alpha.15) (2020-07-27)
+
+
+### Features
+
+* add components/hasura ([75a2f9a](https://github.com/SocialGouv/kosko-charts/commit/75a2f9aa5646dc962abfc1dbf5605b2fb257f452))
+
+# [2.0.0-alpha.14](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.13...v2.0.0-alpha.14) (2020-07-27)
+
+
+### Bug Fixes
+
+* **createDeployment:** handle request/limits ([25f7ae5](https://github.com/SocialGouv/kosko-charts/commit/25f7ae5112688ec86065f2940466be7708d0d626))
+
+# [2.0.0-alpha.13](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.12...v2.0.0-alpha.13) (2020-07-27)
+
+
+### Bug Fixes
+
+* create-db-job name ([2236bc6](https://github.com/SocialGouv/kosko-charts/commit/2236bc6acd613412597a1176bd94f358838a4424))
+
+# [2.0.0-alpha.12](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-alpha.11...v2.0.0-alpha.12) (2020-07-27)
+
+
+### Bug Fixes
+
+* **app:** simplify metadata typing to allow more cases ([0ba7acd](https://github.com/SocialGouv/kosko-charts/commit/0ba7acd2cd71a04b8137c8ba8f6869c56c2f5ad3))
+* **deps:** pin io-ts to 2.2.6 ([2caebbc](https://github.com/SocialGouv/kosko-charts/commit/2caebbc0a725cf7a96fd62c4730ffd3c74bba57a))
+* **release:** add template package.json in release commit ([584263d](https://github.com/SocialGouv/kosko-charts/commit/584263da191a1d5e578727443643b826e32f5912))
+* **release:** missing v before version number ([eea00a4](https://github.com/SocialGouv/kosko-charts/commit/eea00a4d3aee5755f23ec4b1cad5f0f5a6def9e1))
+* **release:** sed the doc ([9d252f4](https://github.com/SocialGouv/kosko-charts/commit/9d252f4b15291e77a171f83efd6706af35c7f452))
+* **release:** sed the doc (2) ([10db9b8](https://github.com/SocialGouv/kosko-charts/commit/10db9b8467a701d43e7f27573a3a5af505c0b667))
+* **release:** sed the doc (3) ([5ffa54e](https://github.com/SocialGouv/kosko-charts/commit/5ffa54e4d000ad05c2916c8b65f83b5de4a463ca))
+* **release:** sed the doc (4) ([0734af1](https://github.com/SocialGouv/kosko-charts/commit/0734af1edc439045d9e69170888d04ea28ab57d6))
+* **release:** sed the doc (5) ([799539e](https://github.com/SocialGouv/kosko-charts/commit/799539e932ec2afef6a09d5bdc271ca769a8256d))
+* **release:** sed the doc (6) ([b0366cc](https://github.com/SocialGouv/kosko-charts/commit/b0366cc9678c2cf84f1c22d604f067653e7dd548))
+* **utils:** committed conflicts ([ed3ac80](https://github.com/SocialGouv/kosko-charts/commit/ed3ac80fd930eb93b675a072bd213d6c804abeab))
+* generation doc ([b5f2a44](https://github.com/SocialGouv/kosko-charts/commit/b5f2a443f34f739e7fa1b9c044c2ee10c79b2d41))
+* **release:** update string to replace in README.md ([413a67f](https://github.com/SocialGouv/kosko-charts/commit/413a67f38032742c1debb6a346cae803cf058865))
+
+
+### Code Refactoring
+
+* **e2e:** run e2e tests on real world templates ([e975403](https://github.com/SocialGouv/kosko-charts/commit/e9754036234e49d15b4914c68ba58c7299ca6f04)), closes [SocialGouv/kosko-charts#2](https://github.com/SocialGouv/kosko-charts/issues/2) [SocialGouv/kosko-charts#2](https://github.com/SocialGouv/kosko-charts/issues/2)
+
+
+### Features
+
+* more utils ([b26991c](https://github.com/SocialGouv/kosko-charts/commit/b26991ccdcb64e41e570037110490452a6285466))
+* **utils:** add helperz ([#65](https://github.com/SocialGouv/kosko-charts/issues/65)) ([787a882](https://github.com/SocialGouv/kosko-charts/commit/787a882dcfc48c4c05fbbfead66ed37022c9ecce))
+* components/azure-db ([#62](https://github.com/SocialGouv/kosko-charts/issues/62)) ([90de6d4](https://github.com/SocialGouv/kosko-charts/commit/90de6d406d7a2c6d8fb0fa702df0310533c88401))
+* update to io-ts 2.2.7 ([400704b](https://github.com/SocialGouv/kosko-charts/commit/400704b78b2637f4cee0a8a9d929b0c70c568b91))
+* **app:** add default livenessProbe and readinessProbe ([03aeba9](https://github.com/SocialGouv/kosko-charts/commit/03aeba926ad1ff8be5284178a70ea8b582c9e168))
+* **app:** add startupProbe ([ec6e468](https://github.com/SocialGouv/kosko-charts/commit/ec6e4680de04d57a730ef45c5ce5f81366535066))
+* **app:** boost default limits ([95c139c](https://github.com/SocialGouv/kosko-charts/commit/95c139cef30c9045af0a6a12a33fcd45af7f844b))
+* **app:** pass params annotations down to pods ([a154598](https://github.com/SocialGouv/kosko-charts/commit/a154598be9f6a486d7757f0fe62242adcf9b50b1))
+* **app:** target our cluster v2 ([fac1347](https://github.com/SocialGouv/kosko-charts/commit/fac13475823ba04c9814ad51e60ea1527104946c))
+* **templates:** update default scripts ([d255d8b](https://github.com/SocialGouv/kosko-charts/commit/d255d8b981d80784aa63806dbe862c26282d2c10))
+* **types:** add NamedComponentEnvironment ([101aa32](https://github.com/SocialGouv/kosko-charts/commit/101aa32401481df724371a196293cb4aa68c4725))
+* **utils:** add public merge function ([fbd0f82](https://github.com/SocialGouv/kosko-charts/commit/fbd0f82c31bb05fe504316b03cc3240dfa8e199b))
+
+
+### BREAKING CHANGES
+
+* **e2e:** we are introduction the `templates` folder
+
+This will change how we scaffold a deployment folder
+We should now be using
+
+```diff
+
 # [2.0.0-beta.17](https://github.com/SocialGouv/kosko-charts/compare/v2.0.0-beta.16...v2.0.0-beta.17) (2020-07-02)